backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
asgiref==3.9.1
async-timeout==5.0.1
attrs==25.3.0
cachetools==5.5.2
certifi==2025.7.14
cffi==1.17.1
channels==4.2.2
charset-normalizer==3.4.2
cryptography==45.0.5
defusedxml==0.7.1
Django==4.2.23
django-anymail==13.0
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
djoser==2.3.3
frozenlist==1.7.0
google-auth==2.40.3
idna==3.10
multidict==6.6.3
oauthlib==3.3.1
pillow==11.3.0
propcache==0.3.2
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pyfcm==2.0.8
PyJWT==2.10.1
python3-openid==3.2.0
qrcode==8.2
reportlab==4.4.3
requests==2.32.4
requests-oauthlib==2.0.0
rsa==4.9.1
social-auth-app-django==5.4.3
social-auth-core==4.7.0
sqlparse==0.5.3
typing_extensions==4.14.1
urllib3==2.5.0
yarl==1.20.1
